refactoring components and moving the Speaker comp and its relevant dependencies into a new dir
