+++ pages/index.js
@@ -1,20 +1,5 @@
 import { data } from "../SpeakerData";
-
-const Sessions = ({ sessions }) => {
-  return (
-    <div className="sessionBox card h-250">
-      <Session title={sessions[0].title} room={sessions[0].room.name} />
-    </div>
-  );
-};
-
-const Session = ({ title, room }) => {
-  return (
-    <span className="session w-100">
-      {title} <strong>Room: {room}</strong>
-    </span>
-  );
-};
+import Speaker from "../src/components/Speaker";
 
 const IndexPage = () => {
   return (
@@ -31,35 +16,7 @@ const IndexPage = () => {
             company,
             sessions,
           } = speaker;
-          return (
-            <div
-              key={id}
-              className="col-xs-12 col-sm-12 col-md-6 col-lg-4 col-sm-12 col-xs-12"
-            >
-              <div className="card card-height p-4 mt-4">
-                <div className="speaker-img d-flex flex-row justify-content-center align-items-center h-300">
-                  <img
-                    className="contain-fit"
-                    src={`/images/speaker-${id}.jpg`}
-                    width="300"
-                  />
-                </div>
-                <div className="speaker-info">
-                  <div className="d-flex justify-content-between mb-3">
-                    <h3>
-                      {first} {last}
-                    </h3>
-                  </div>
-                  <div>
-                    <p>
-                      {bio} {company} {twitterHandle} {favorite}
-                    </p>
-                  </div>
-                </div>
-                <Sessions sessions={sessions} />
-              </div>
-            </div>
-          );
+          return <Speaker key={speaker.id} speaker={speaker} />;
         })}
       </div>
     </div>
